requirements.txt
langgraph==1.0.10
langchain-core==1.2.16
langchain-groq==1.1.2
biopython==1.86
chembl-webresource-client==0.10.9
python-dotenv==1.2.2
pydantic==2.12.5